pyproject.toml
[project]
name = "abp-eng-dados"
version = "0.1.0"
description = ""
authors = [
    {name = "pauloFrancisconi",email = "pauloffrancisconi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faker (>=37.3.0,<38.0.0)",
    "pyodbc (>=5.2.0,<6.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
